Split test_harness.rs (787 lines) into subdirectory modules
Split the monolithic test_harness.rs into four focused files under
test_harness/:
- mod.rs (114 lines): struct definition + constructors
- setup.rs (319 lines): road, zone, building, service, utility builders
- spawning.rs (124 lines): citizen spawning methods
- queries.rs (178 lines): query + simulation tick methods
- assertions.rs (108 lines): assertion helpers

All pub items remain accessible via `crate::test_harness::TestCity`.

Part of #1389
